spec: unflake some net specs
