+++ spec/api-net-spec.js
@@ -61,187 +61,6 @@ describe('net module', () => {
     })
   })
 
-  describe('HTTP basics', () => {
-    it('should be able to issue a basic GET request', (done) => {
-      const requestUrl = '/requestUrl'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.method).to.equal('GET')
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request(`${server.url}${requestUrl}`)
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should be able to issue a basic POST request', (done) => {
-      const requestUrl = '/requestUrl'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.method).to.equal('POST')
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        method: 'POST',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should fetch correct data in a GET request', (done) => {
-      const requestUrl = '/requestUrl'
-      const bodyData = 'Hello World!'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.method).to.equal('GET')
-            response.write(bodyData)
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request(`${server.url}${requestUrl}`)
-      urlRequest.on('response', (response) => {
-        let expectedBodyData = ''
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-          expectedBodyData += chunk.toString()
-        })
-        response.on('end', () => {
-          expect(expectedBodyData).to.equal(bodyData)
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should post the correct data in a POST request', (done) => {
-      const requestUrl = '/requestUrl'
-      const bodyData = 'Hello World!'
-      server.on('request', (request, response) => {
-        let postedBodyData = ''
-        switch (request.url) {
-          case requestUrl:
-            expect(request.method).to.equal('POST')
-            request.on('data', (chunk) => {
-              postedBodyData += chunk.toString()
-            })
-            request.on('end', () => {
-              expect(postedBodyData).to.equal(bodyData)
-              response.end()
-            })
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        method: 'POST',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {})
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.write(bodyData)
-      urlRequest.end()
-    })
-
-    it('should support chunked encoding', (done) => {
-      const requestUrl = '/requestUrl'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.chunkedEncoding = true
-            expect(request.method).to.equal('POST')
-            expect(request.headers['transfer-encoding']).to.equal('chunked')
-            expect(request.headers['content-length']).to.be.undefined()
-            request.on('data', (chunk) => {
-              response.write(chunk)
-            })
-            request.on('end', (chunk) => {
-              response.end(chunk)
-            })
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        method: 'POST',
-        url: `${server.url}${requestUrl}`
-      })
-
-      let chunkIndex = 0
-      const chunkCount = 100
-      const sentChunks = []
-      const receivedChunks = []
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-          receivedChunks.push(chunk)
-        })
-        response.on('end', () => {
-          const sentData = Buffer.concat(sentChunks)
-          const receivedData = Buffer.concat(receivedChunks)
-          expect(sentData.toString()).to.equal(receivedData.toString())
-          expect(chunkIndex).to.be.equal(chunkCount)
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.chunkedEncoding = true
-      while (chunkIndex < chunkCount) {
-        chunkIndex += 1
-        const chunk = randomBuffer(kOneKiloByte)
-        sentChunks.push(chunk)
-        expect(urlRequest.write(chunk)).to.be.true()
-      }
-      urlRequest.end()
-    })
-  })
-
   describe('ClientRequest API', () => {
     afterEach(() => {
       session.defaultSession.webRequest.onBeforeRequest(null)
